Removed unused code, updated tests, renamed file
