+++ game/kegeland/state-management/game/gameMode.ts
@@ -1,5 +1,5 @@
 export enum GameMode {
   OneControl = 'One control',
   MultiControl = 'Multiple controls',
-  SensorDataTestControl = 'SensorDataTest Controls',
+  SensorDataTestControl = 'SensorDataTest controls',
 }
